refactor: remove legacy chat model admin panel (#26452)
