+++ site/src/pages/AgentsPage/components/ChatModelAdminPanel/ProviderField.tsx
@@ -1,0 +1,35 @@
+import type { FC, ReactNode } from "react";
+
+interface ProviderFieldProps {
+	label: string;
+	htmlFor?: string;
+	required?: boolean;
+	description?: string;
+	children: ReactNode;
+}
+
+export const ProviderField: FC<ProviderFieldProps> = ({
+	label,
+	htmlFor,
+	required,
+	description,
+	children,
+}) => (
+	<div className="grid gap-1.5">
+		<div className="flex items-baseline gap-1.5">
+			<label
+				htmlFor={htmlFor}
+				className="text-sm font-medium text-content-primary"
+			>
+				{label}
+			</label>
+			{required && (
+				<span className="text-xs font-bold text-content-destructive">*</span>
+			)}
+		</div>
+		{description && (
+			<p className="m-0 text-xs text-content-secondary">{description}</p>
+		)}
+		{children}
+	</div>
+);
